permission function changed LOGIN

+++ app/Http/Controllers/Api/V1/User/UserController.php
@@ -35,28 +35,11 @@ class UserController extends Controller
         $this->middleware('permission:users-delete', ['only' => ['destroy']]);*/
         
         $this->middleware(function ($request, $next) {
-            if(auth()->user()->user_type_id=='1') {
-                $this->top_most_parent_id = auth()->user()->id;
-            }
-            else if(auth()->user()->user_type_id=='2')
-            {
-                $this->top_most_parent_id = auth()->user()->id;
-            } else {
-                $this->top_most_parent_id = auth()->user()->top_most_parent_id;
-            }
+            $this->top_most_parent_id = auth()->user()->top_most_parent_id;
             return $next($request);
         });
     }
 
-    /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
-    public function index()
-    {
-        //
-    }
     public function users(Request $request)
     {
         try {
@@ -179,12 +162,22 @@ class UserController extends Controller
                 $password = Str::random(12);
             }
 
+            //for role set
+            if($request->user_type_id==6)
+            {
+                $roleInfo = getRoleInfo($this->top_most_parent_id, 'Patient');
+            }
+            else
+            {
+                $roleInfo = Role::where('id',$request->role_id)->first();
+            }
+
             $user = new User;
             $user->unique_id = generateRandomNumber();
             $user->branch_id = getBranchId();
             $user->custom_unique_id = $request->custom_unique_id;
             $user->user_type_id = $request->user_type_id;
-            $user->role_id = $request->role_id;
+            $user->role_id = $roleInfo->id;
             $user->company_type_id = (!empty($request->company_type_id)) ? json_encode($request->company_type_id) : $userInfo->company_type_id;
             $user->category_id = (!empty($request->category_id)) ? $request->category_id : $userInfo->category_id;
             $user->top_most_parent_id = $this->top_most_parent_id;
@@ -224,9 +217,9 @@ class UserController extends Controller
             $user->documents = json_encode($request->documents);
             $user->entry_mode =  (!empty($request->entry_mode)) ? $request->entry_mode :'Web';
             $user->save();
-            if(!empty($request->input('role_id')))
+            if($roleInfo)
             {
-                $role = Role::where('id',$request->role_id)->first();
+                $role = $roleInfo;
                 $user->assignRole($role->name);
             }
             if($request->user_type_id == '6'){
@@ -413,9 +406,9 @@ class UserController extends Controller
     {
         try {
             
-            $checkId= User::where('id',$user->id)->where('top_most_parent_id',$this->top_most_parent_id)->first();
+            $checkId= User::where('id', $user->id)->where('top_most_parent_id', $this->top_most_parent_id)->first();
             if (!is_object($checkId)) {
-                return prepareResult(falsuser,getLangByLabelGroups('UserValidation','id_not_found'), [],'404');
+                return prepareResult(false,getLangByLabelGroups('UserValidation','id_not_found'), [],'404');
             }
             $userShow = User::where('id',$user->id)->with('TopMostParent:id,user_type_id,name,email','UserType:id,name','CategoryMaster:id,created_by,name','Department:id,name','Country:id,name','weeklyHours','branch','persons.Country','PatientInformation','branch:id,name,email,contact_number')->first();
             return prepareResult(true,'User View' ,$userShow, '200');
@@ -469,9 +462,19 @@ class UserController extends Controller
             if (!is_object($checkId)) {
                 return prepareResult(false, getLangByLabelGroups('UserValidation','id_not_found'), [],'404');
             }
+
+            //for role set
+            if($request->user_type_id==6)
+            {
+                $roleInfo = getRoleInfo($this->top_most_parent_id, 'Patient');
+            }
+            else
+            {
+                $roleInfo = Role::where('id',$request->role_id)->first();
+            }
             
             $user->user_type_id = $request->user_type_id;
-            $user->role_id = $request->role_id;
+            $user->role_id = $roleInfo->id;
             $user->category_id = (!empty($request->category_id)) ? $request->category_id : $userInfo->category_id;
             $user->govt_id = $request->govt_id;
             $user->dept_id = $request->dept_id;
@@ -503,10 +506,11 @@ class UserController extends Controller
             $user->entry_mode =  (!empty($request->entry_mode)) ? $request->entry_mode :'Web';
             $user->documents = json_encode($request->documents);
             $user->save();
-            \DB::table('model_has_roles')->where('model_id',$user->id)->delete();
-            if(!empty($request->input('role_id')))
+            
+            if($roleInfo)
             {
-                $role = Role::where('id',$request->role_id)->first();
+                \DB::table('model_has_roles')->where('model_id',$user->id)->delete();
+                $role = $roleInfo;
                 $user->assignRole($role->name);
             }
 
@@ -612,15 +616,8 @@ class UserController extends Controller
                         $personalInfo->save() ;
                         /*-----Create Account /Entry in user table*/
                         if($is_user == true) {
-                            if(auth()->user()->user_type_id=='1'){
-                                $top_most_parent_id = auth()->user()->id;
-                            }
-                            elseif(auth()->user()->user_type_id=='2')
-                            {
-                                $top_most_parent_id = auth()->user()->id;
-                            } else {
-                                $top_most_parent_id = auth()->user()->top_most_parent_id;
-                            }
+                            $top_most_parent_id = auth()->user()->top_most_parent_id;
+
                             $checkAlreadyUser = User::where('email',@$value['email'])->first();
                             if(empty($checkAlreadyUser)) {
                                 if(!empty($getUser)){
@@ -677,123 +674,6 @@ class UserController extends Controller
         }
     }
 
-    /**
-     * Remove the specified resource from storage.
-     *
-     * @param  \App\Models\User  $user
-     * @return \Illuminate\Http\Response
-     */
-
-     public function importPatient(Request $request,User $user){
-        DB::beginTransaction();
-        try {
-            $userInfo = getUser();
-            $validator = Validator::make($request->all(),[
-                 'file'     => 'required|array|max:20000|min:1'
-            ]);
-            if ($validator->fails()) {
-                return prepareResult(false,$validator->errors()->first(),[], '422'); 
-            }
-            
-            if($request->user_type_id == '6' || $request->user_type_id =='3'){
-                $validator = Validator::make($request->all(),[
-                    'personal_number' => 'required|digits:12|unique:users,personal_number,'.$user->id, 
-                ]);
-                if ($validator->fails()) {
-                    return prepareResult(false,$validator->errors()->first(),[], '422'); 
-                }
-
-            }
-            
-            $checkId = User::where('id',$user->id)->where('top_most_parent_id',$this->top_most_parent_id)->first();
-            if (!is_object($checkId)) {
-                return prepareResult(false, getLangByLabelGroups('UserValidation','id_not_found'), [],'404');
-            }
-            
-            $user->user_type_id = $request->user_type_id;
-            $user->role_id = $request->role_id;
-            $user->category_id = (!empty($request->category_id)) ? $request->category_id : $userInfo->category_id;
-            $user->govt_id = $request->govt_id;
-            $user->dept_id = $request->dept_id;
-            $user->country_id = $request->country_id;
-            $user->city = $request->city;
-            $user->postal_area = $request->postal_area;
-            $user->name = $request->name;
-            $user->contact_number = $request->contact_number;
-            $user->gender = $request->gender;
-            $user->personal_number = $request->personal_number;
-            $user->organization_number = $request->organization_number;
-            $user->patient_type_id = (!empty($request->patient_type_id)) ? json_encode($request->patient_type_id) : null;
-            $user->zipcode = $request->zipcode;
-            $user->full_address = $request->full_address;
-            $user->joining_date = $request->joining_date;
-            $user->establishment_year = $request->establishment_year;
-            $user->user_color = $request->user_color;
-            $user->disease_description = $request->disease_description;
-            $user->is_substitute = ($request->is_substitute) ? 1:0 ;
-            $user->is_regular = ($request->is_regular) ? 1:0 ;
-            $user->is_seasonal = ($request->is_seasonal) ? 1:0 ;
-            $user->is_file_required = ($request->is_file_required) ? 1:0 ;
-            $user->is_secret = ($request->is_secret) ? 1:0 ;
-            $user->step_one = (!empty($request->step_one)) ? $request->step_one:0 ;
-            $user->step_two = (!empty($request->step_two)) ? $request->step_two:0 ;
-            $user->step_three = (!empty($request->step_three)) ? $request->step_three:0 ;
-            $user->step_four = (!empty($request->step_four)) ? $request->step_four:0 ;
-            $user->step_five = (!empty($request->step_five)) ? $request->step_five:0 ;
-            $user->entry_mode =  (!empty($request->entry_mode)) ? $request->entry_mode :'Web';
-            $user->documents = json_encode($request->documents);
-            $user->save();
-            \DB::table('model_has_roles')->where('model_id',$user->id)->delete();
-            if(!empty($request->input('role_id')))
-            {
-                $role = Role::where('id',$request->role_id)->first();
-                $user->assignRole($role->name);
-            }
-
-
-            if($request->user_type_id == '6'){
-                $checkPatientAlready = PatientInformation::where('patient_id',$user->id)->first();
-                if(is_object($checkPatientAlready)){
-                    $patientInfo =  PatientInformation::find($checkPatientAlready->id);
-                } else{
-                    $patientInfo = new PatientInformation;
-                }
-                
-                $patientInfo->patient_id = $user->id;
-                $patientInfo->institute_name = $request->institute_name;
-                $patientInfo->institute_contact_number = $request->institute_contact_number;
-                $patientInfo->institute_full_address = $request->institute_full_address;
-                $patientInfo->classes_from = $request->classes_from;
-                $patientInfo->classes_to = $request->classes_to;
-                $patientInfo->company_name = $request->company_name;
-                $patientInfo->company_contact_number = $request->company_contact_number;
-                $patientInfo->company_full_address = $request->company_full_address;
-                $patientInfo->from_timing = $request->from_timing;
-                $patientInfo->to_timing = $request->to_timing;
-                $patientInfo->special_information = $request->special_information;
-                $patientInfo->aids = $request->aids;
-                $patientInfo->another_activity = $request->another_activity;
-                $patientInfo->another_activity_name = $request->another_activity_name;
-                $patientInfo->activitys_contact_number = $request->activitys_contact_number;
-                $patientInfo->activitys_full_address = $request->activitys_full_address;
-                $patientInfo->week_days = json_encode($request->week_days);
-                $patientInfo->issuer_name = $request->issuer_name;
-                $patientInfo->number_of_hours = $request->number_of_hours;
-                $patientInfo->period = $request->period;
-                $patientInfo->save();
-              
-            }
-            DB::commit();
-            return prepareResult(true,getLangByLabelGroups('UserValidation','update'),$user, '200');
-                
-        }
-        catch(Exception $exception) {
-             \Log::error($exception);
-            DB::rollback();
-            return prepareResult(false, $exception->getMessage(),[],'500');
-            
-        }
-    }
     public function destroy(User $user)
     {
         try {
